refactor(transport): the submit machinery moves out, and the stranded doc moves home
The develop merge put `packages/agent-transport-http/src/routes.ts` back over
the 300-line ceiling (307): it now carries the admission gate AND the claim
machinery AND the route list. The split follows the seam the review rounds
already carved — `packages/agent-transport-http/src/submit-route.ts` takes `admitTurn` and `submitHandler` (the
concurrency-admission decision and the claim lifecycle), `packages/agent-transport-http/src/routes.ts` stays the
list of endpoints it reads as. 190 + 137 lines, no function over 50.

And review caught this file doing the thing it keeps writing about: the
`release` JSDoc stayed on `wireRelay` when the split moved the parameter to
`relayTurn`. It sits on `relayTurn` now, with the catch noted in it.
